+++ familyhelper-ui-view/packages/webapp-mobile/src/views/nodes/comvisual/hyperscript/placeholder/index.ts
@@ -1,6 +1,6 @@
 import { type Component, defineComponent, h } from 'vue'
 
-import BorderLayoutPanel from '@/components/comvisual/layout/borderLayoutPanel/BorderLayoutPanel.vue'
+import RootBorderLayoutPanel from '@/components/comvisual/layout/rootBorderLayoutPanel/RootBorderLayoutPanel.vue'
 import PlaceholderPanel from '@/components/comvisual/layout/placeholderPanel/PlaceholderPanel.vue'
 
 import styles from './index.module.css'
@@ -12,8 +12,11 @@ export function placeholder(placeholder: string): Component {
       return () =>
         h('div', { class: styles['placeholder-container'] }, [
           h(
-            BorderLayoutPanel,
-            {},
+            RootBorderLayoutPanel,
+            {
+              initialDockStatus: 4,
+              initialFloatyY: -200,
+            },
             {
               default: () => h(PlaceholderPanel, { text: placeholder }),
             },